refactor(javm-cap): split cap.rs into cap/ directory (one file per variant)
The five Cap variants now live as siblings under a `cap/` directory:

```
src/
├── lib.rs
├── abi.rs
├── cache.rs       — CacheDirectory + CapRef + CapHashOrRef + CacheError
├── cap/
│   ├── mod.rs     — Cap enum, CapHash, CapKind, TypeCap, NUM_REGS, MAX_*, constructors
│   ├── cnode.rs   — CNodeCap, MAX_CNODE_SLOTS
│   ├── data.rs    — DataCap, DataContent, PAGE_SIZE
│   ├── image.rs   — ImageCap, EndpointDef, MemoryMapping, ImageSlotEntry, image_cap()
│   ├── instance.rs — InstanceCap, RwOverlay
│   └── page.rs    — PageSlot, PageBytes, PageRef
├── error.rs
├── hash.rs
├── image.rs       — Image manifest (SCALE form) — distinct from cap::image
├── slot.rs
└── wire.rs
```

`image_cap.rs` becomes `cap::image::ImageCap`, paired against the
manifest `crate::image::Image`. The naming asymmetry that lib.rs was
papering over (`EndpointDef as ImageEndpointDef`, `MemoryMapping as
ImageMemoryMapping`) goes away — the cap-form types are at
`cap::image::{EndpointDef, MemoryMapping}` and the manifest types
at `image::{EndpointDef, MemoryMapping}`.

Internal imports rewritten throughout javm-cap. External crates that
named the qualified paths updated:
- `nub-arch-local::lib` — `javm_cap::image_cap::ImageCap` → `javm_cap::cap::image::ImageCap`
- `nub-arch-x86::call_loop` — `javm_cap::instance::InstanceCap` → `javm_cap::cap::instance::InstanceCap`
- `javm::ecall` — `javm_cap::data::alloc_page_aligned_zeroed` /
  `javm_cap::page::PageSlot` → `javm_cap::cap::data::*` /
  `javm_cap::cap::page::*`

The canonical workspace paths (`javm_cap::Cap`, `javm_cap::ImageCap`,
etc.) are unchanged — `lib.rs` still re-exports them at the crate
root.
