+++ rust/javm/src/ecall.rs
@@ -705,7 +705,7 @@ impl<K: KernelAssist> Vm<K> {
         // Quota is debited by the padded length — the kernel owns
         // a full page-aligned allocation regardless of caller's slice
         // length, so callers pay for what they store.
-        let mut inline = javm_cap::data::alloc_page_aligned_zeroed(bytes.len());
+        let mut inline = javm_cap::cap::data::alloc_page_aligned_zeroed(bytes.len());
         inline[..bytes.len()].copy_from_slice(&bytes);
         let debit = inline.len() as u64;
         let quota = self.kernel_assist.storage_quota_get(quota_id);
@@ -877,7 +877,7 @@ fn data_cap_prefix(data: &DataCap, len: usize) -> Vec<u8> {
                     break;
                 }
                 let end = (start + page_size).min(actual_len);
-                if let javm_cap::page::PageSlot::Loaded(page_ref) = page {
+                if let javm_cap::cap::page::PageSlot::Loaded(page_ref) = page {
                     let page_bytes = &page_ref.bytes;
                     out[start..end].copy_from_slice(&page_bytes[..end - start]);
                 }
